add tabs to SavedViewsModal
