+++ src/components/modals/SavedViewsModal.svelte
@@ -1,100 +1,30 @@
 <svelte:options immutable={true} />
 
 <script lang="ts">
-  import type { ICellRendererParams, ValueFormatterParams } from 'ag-grid-community';
   import { createEventDispatcher } from 'svelte';
+  import { user } from '../../stores/app';
   import { initializeView, view, views } from '../../stores/views';
-  import type { DataGridColumnDef } from '../../types/data-grid';
   import type { View } from '../../types/view';
   import effects from '../../utilities/effects';
   import { setQueryParam } from '../../utilities/generic';
-  import BulkActionDataGrid from '../ui/DataGrid/BulkActionDataGrid.svelte';
-  import DataGridActions from '../ui/DataGrid/DataGridActions.svelte';
+  import Tab from '../ui/Tabs/Tab.svelte';
+  import TabPanel from '../ui/Tabs/TabPanel.svelte';
+  import Tabs from '../ui/Tabs/Tabs.svelte';
+  import ViewsTable from '../view/ViewsTable.svelte';
   import Modal from './Modal.svelte';
   import ModalContent from './ModalContent.svelte';
-  import ModalFooter from './ModalFooter.svelte';
   import ModalHeader from './ModalHeader.svelte';
 
   export let height: number | string = 150;
   export let width: number | string = 380;
 
-  type CellRendererParams = {
-    deleteView: (view: View) => void;
-    openView: (view: View) => void;
-  };
-  type ViewCellRendererParams = ICellRendererParams<View> & CellRendererParams;
-
   const dispatch = createEventDispatcher();
-  const columnDefs: DataGridColumnDef[] = [
-    {
-      field: 'id',
-      filter: 'number',
-      headerName: 'ID',
-      resizable: true,
-      sortable: true,
-      suppressAutoSize: true,
-      suppressSizeToFit: true,
-      width: 60,
-    },
-    { field: 'name', filter: 'text', headerName: 'Name', resizable: true, sortable: true },
-    { field: 'owner', filter: 'text', headerName: 'Owner', resizable: true, sortable: true },
-    {
-      field: 'updated_at',
-      filter: 'text',
-      headerName: 'Last Edited',
-      resizable: true,
-      sortable: true,
-      valueFormatter: ({ value: updatedAt }: ValueFormatterParams<View, string>) => {
-        const updatedAtDate = new Date(updatedAt);
-        updatedAtDate.setMilliseconds(0);
 
-        return updatedAtDate.toISOString().replace(/.\d+Z$/g, 'Z');
-      },
-    },
-    {
-      cellClass: 'action-cell-container',
-      cellRenderer: (params: ViewCellRendererParams) => {
-        const isEditable = params.data.owner !== 'system';
-        const actionsDiv = document.createElement('div');
-        actionsDiv.className = 'actions-cell';
-        new DataGridActions({
-          props: {
-            ...(isEditable
-              ? {
-                  deleteCallback: params.deleteView,
-                  deleteTooltip: {
-                    content: 'Delete View',
-                    placement: 'bottom',
-                  },
-                }
-              : {}),
-            rowData: params.data,
-            viewCallback: params.openView,
-            viewTooltip: {
-              content: 'Open View',
-              placement: 'bottom',
-            },
-          },
-          target: actionsDiv,
-        });
+  let userViews: View[] = [];
 
-        return actionsDiv;
-      },
-      cellRendererParams: {
-        deleteView,
-        openView,
-      } as CellRendererParams,
-      field: 'actions',
-      headerName: '',
-      resizable: false,
-      sortable: false,
-      suppressAutoSize: true,
-      suppressSizeToFit: true,
-      width: 55,
-    },
-  ];
+  $: userViews = $views.filter((view: View) => view.owner === $user.id);
 
-  async function deleteView({ id: viewId }: View) {
+  async function deleteView({ detail: viewId }: CustomEvent<number>) {
     const success = await effects.deleteView(viewId);
 
     if (success) {
@@ -104,12 +34,11 @@
     }
   }
 
-  async function deleteViews({ detail: views }: CustomEvent<View[]>) {
-    const ids = views.map(({ id }) => id);
-    const success = await effects.deleteViews(ids);
+  async function deleteViews({ detail: viewIds }: CustomEvent<number[]>) {
+    const success = await effects.deleteViews(viewIds);
 
     if (success) {
-      if (ids.includes($view.id)) {
+      if (viewIds.includes($view.id)) {
         goToNextView();
       }
     }
@@ -121,7 +50,7 @@
     setQueryParam('viewId', `${nextView.id}`);
   }
 
-  async function openView({ id: viewId }: View) {
+  async function openView({ detail: viewId }: CustomEvent<number>) {
     const query = new URLSearchParams(`?viewId=${viewId}`);
     const newView = await effects.getView(query);
 
@@ -137,18 +66,45 @@
 
 <Modal {height} {width}>
   <ModalHeader on:close>Saved Views</ModalHeader>
-  <ModalContent>
-    <BulkActionDataGrid
-      {columnDefs}
-      isRowSelectable={rowData => rowData.data.owner !== 'system'}
-      items={$views}
-      pluralItemDisplayText="Views"
-      singleItemDisplayText="View"
-      on:bulkDeleteItems={deleteViews}
-      on:rowDoubleClicked={event => openView(event.detail)}
-    />
+  <ModalContent style="padding:0">
+    <Tabs class="view-tabs" tabListClassName="view-tabs-list">
+      <svelte:fragment slot="tab-list">
+        <Tab class="view-tab">All Views</Tab>
+        <Tab class="view-tab">My Views</Tab>
+      </svelte:fragment>
+      <TabPanel>
+        <ViewsTable views={$views} on:delete-view={deleteView} on:delete-views={deleteViews} on:open-view={openView} />
+      </TabPanel>
+      <TabPanel>
+        <ViewsTable
+          views={userViews}
+          on:delete-view={deleteView}
+          on:delete-views={deleteViews}
+          on:open-view={openView}
+        />
+      </TabPanel>
+    </Tabs>
   </ModalContent>
-  <ModalFooter>
-    <button class="st-button" on:click={() => dispatch('close')}> Close </button>
-  </ModalFooter>
 </Modal>
+
+<style>
+  :global(.view-tabs) {
+    padding: 0 1rem 1rem;
+  }
+
+  :global(.view-tab) {
+    --tab-background-color: var(--st-white);
+    --tab-hover-background-color: var(--st-white);
+    --tab-hover-text-color: var(--st-gray-70);
+    --tab-padding: 0px;
+    --tab-selected-background-color: var(--st-white);
+    --tab-text-color: var(--st-gray-50);
+  }
+
+  :global(.view-tabs-list) {
+    --tab-list-background-color: var(--st-white);
+    --tab-list-gap: 8px;
+    border-bottom: 1px solid var(--st-gray-15);
+    width: 100%;
+  }
+</style>
